refactor(pipeline): break 3 orphan import cycles
Phase 12g-1 of the acyclic-dependencies burn-down. Break three
localized 2-node import cycles so the pipeline graph ratchets from
10 to 7 frozen cycles:

- Init: drop the InitPhase -> InitPhaseFactory re-export; consumers
  import createInitPhase/INIT_STEP from the factory directly, and
  InitPhase becomes a default export (single-export rule).
- ApiDirectScrape: extract ApiDirectScrapeResult/ApiDirectScrapeFn
  into ApiDirectScrapeTypes.ts so Actions and Phase both depend on
  the leaf instead of each other. Phase re-exports for consumers.
- Registry/Config: source the seeder types from
  PipelineBankConfigTypes.js, not the PipelineBankConfig.js value.

Pure import-graph rewiring + one type-extraction leaf; no behavior
change. Baseline re-frozen to 7. type-check, lint, cycles, and the
affected unit tests all pass.
