+++ src/Scrapers/Pipeline/Phases/ApiDirectScrape/ApiDirectScrapePhase.ts
@@ -15,7 +15,7 @@
 
 import { logForensicAudit } from '../../Mediator/Scrape/ForensicAuditAction.js';
 import { BasePhase } from '../../Types/BasePhase.js';
-import { type Option, some } from '../../Types/Option.js';
+import { some } from '../../Types/Option.js';
 import type {
   IActionContext,
   IPipelineContext,
@@ -24,22 +24,10 @@ import type {
 import type { Procedure } from '../../Types/Procedure.js';
 import { succeed } from '../../Types/Procedure.js';
 import { buildGenericHeadlessScrape } from './ApiDirectScrapeActions.js';
+import type { ApiDirectScrapeFn } from './ApiDirectScrapeTypes.js';
 import type { IApiDirectScrapeShape } from './IApiDirectScrapeShape.js';
 
-/**
- * Action-context payload returned by the shape-driven scrape function:
- * the sealed action context augmented with the `scrape` slot that
- * DASHBOARD-style phases would otherwise commit. The intersection is
- * a true subtype of {@link IActionContext}, so this Procedure is
- * directly assignable to `Procedure<IActionContext>` (the shape
- * required by {@link BasePhase.action}) without an unsafe cast.
- */
-export type ApiDirectScrapeResult = IActionContext & {
-  readonly scrape: Option<IScrapeState>;
-};
-
-/** Bound phase action — the shape-driven scrape function. */
-export type ApiDirectScrapeFn = (ctx: IActionContext) => Promise<Procedure<ApiDirectScrapeResult>>;
+export type { ApiDirectScrapeFn, ApiDirectScrapeResult } from './ApiDirectScrapeTypes.js';
 
 /** ApiDirectScrape phase — BasePhase bound to a shape literal. */
 class ApiDirectScrapePhase extends BasePhase {
